ref(feedback): Rename Dialog to Form across feedback APIs
Rename SentryUserFeedbackDialog to SentryUserFeedbackForm as the primary
class. Keep SentryUserFeedbackDialog as a deprecated subclass for backward
compatibility.

Also rename internal APIs to use Form naming consistently:
- IDialogHandler -> IFormHandler
- showDialog -> showForm
- setDialogHandler/getDialogHandler -> setFormHandler/getFormHandler
- AndroidUserFeedbackIDialogHandler -> AndroidUserFeedbackFormHandler

Add deprecated Sentry.showUserFeedbackDialog() overloads that delegate to
the new Sentry.showUserFeedbackForm() methods.
