+++ sentry/src/test/java/io/sentry/SentryFeedbackOptionsTest.kt
@@ -1,6 +1,6 @@
 package io.sentry
 
-import io.sentry.SentryFeedbackOptions.IDialogHandler
+import io.sentry.SentryFeedbackOptions.IFormHandler
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import org.mockito.kotlin.mock
@@ -8,7 +8,7 @@ import org.mockito.kotlin.mock
 class SentryFeedbackOptionsTest {
   @Test
   fun `feedback options is initialized with default values`() {
-    val options = SentryFeedbackOptions(mock<IDialogHandler>())
+    val options = SentryFeedbackOptions(mock<IFormHandler>())
     assertEquals(false, options.isNameRequired)
     assertEquals(true, options.isShowName)
     assertEquals(false, options.isEmailRequired)
@@ -35,7 +35,7 @@ class SentryFeedbackOptionsTest {
   @Test
   fun `feedback options copy constructor`() {
     val options =
-      SentryFeedbackOptions(mock<IDialogHandler>()).apply {
+      SentryFeedbackOptions(mock<IFormHandler>()).apply {
         isNameRequired = true
         isShowName = false
         isEmailRequired = true
@@ -80,6 +80,6 @@ class SentryFeedbackOptionsTest {
     assertEquals(options.onFormClose, optionsCopy.onFormClose)
     assertEquals(options.onSubmitSuccess, optionsCopy.onSubmitSuccess)
     assertEquals(options.onSubmitError, optionsCopy.onSubmitError)
-    assertEquals(options.dialogHandler, optionsCopy.dialogHandler)
+    assertEquals(options.formHandler, optionsCopy.formHandler)
   }
 }
